Refactor functions to logical units/ classes
