+++ bin/main.rb
@@ -7,13 +7,13 @@ require 'csv'
 require 'io/console'
 
 require_relative '../lib/Display.rb'
+require_relative '../lib/Scraper.rb'
 screen_o = Display.new
+scraper_o = Scraper.new
 
-$post_game = ''
-$sub_path = ''
-$index_path = ''
 $choice_input = ''
-$has_index = true
+
+$sub_path = ''
 
 def check_choice
   $choice_input = gets.strip.upcase
@@ -29,66 +29,17 @@ def check_choice
   end
 end
 
-def scraper
-  url = "https://apps.apple.com/us/genre/ios-games/id6014/#{$sub_path}#{$index_path}"
-
-  unparsed_page = HTTParty.get(url)
-  parsed_page = Nokogiri::HTML(unparsed_page)
-
-  games_array = []
-
-  # parse the data
-  parsed_page.css('[id = "content"]').css('[id = "selectedgenre"]').css('[id = "selectedcontent"]').map do |a|
-    $post_game = a.text
-    games_array.push($post_game)
-  end
-
-   # clear and write content to csv
-   File.open('../lib/iOSGames.csv', 'w') {}
-   CSV.open('../lib/iOSGames.csv', 'w') do |csv|
-     csv << games_array
-   end
-   # Print CSV to console
-   CSV.foreach('../lib/iOSGames.csv') do |row|
-     puts row
-   end
-
-  # Pry.start(binding)
-end
-
-def check_index
-  page_index = 1
-  while $has_index
-    # $has_index = false if $post_game.match(/^[[:blank:]]$/)
-    
-    puts "press '>' to goto next indexed Page for Games indexed #{$choice_input}"
-    puts "press '<' to goto previous indexed Page"
-    next_input = STDIN.getch
-
-    case next_input
-      when ?.
-        page_index += 1
-        $index_path = "&page=#{page_index}#page"
-        scraper
-      when ?,
-        page_index -= 1
-        $index_path = "&page=#{page_index}#page"
-        scraper
-    end
   #test
-  $has_index = false if $post_game.size < 500
-  puts "Indexed page is #{page_index}"
-  puts "index path is #{$index_path}"
-  puts "#{$has_index}"
-  puts "my array count is #{$post_game.size}"
-  
-  end
-end
+  # $has_index = false if $post_game.size < 500
+  # puts "Indexed page is #{page_index}"
+  # puts "index path is #{$index_path}"
+  # puts "#{$has_index}"
+  # puts "my array count is #{$post_game.size}"
 
 screen_o.display_intro_screen
 
 check_choice
 
-scraper
+scraper_o.scrape
 
-check_index
+scraper_o.check_index
